requirements.txt
fastapi>=0.116.0
uvicorn[standard]>=0.35.0
python-dotenv>=1.0.1
python-multipart>=0.0.9
requests>=2.32.3
openai>=1.68.2
langchain-core>=1.0.0
langchain-text-splitters>=1.0.0
pymilvus[milvus-lite]>=2.5.4
pypdf>=5.3.1
